requirements.txt
python-binance==0.7.9
request==2.22.0
python-decouple==3.4